third_party/blink/web_tests/http/tests/websocket/server-close-expected.txt

WebSocket: Test server-initiated close.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
Closed
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
